Move to imgui v1.91.9b

+++ imgui/src/fonts/glyph_ranges.rs
@@ -62,7 +62,7 @@ impl FontGlyphRanges {
     // to show for unpaired surrogates) Would be nice to be sure, if so, this
     // should accept `char` (we'd still have to check that the range doesn't
     // fully contain the surrogate range though)
-    pub fn from_slice(slice: &'static [u32]) -> FontGlyphRanges {
+    pub fn from_slice(slice: &'static [u16]) -> FontGlyphRanges {
         assert_eq!(
             slice.len() % 2,
             1,
@@ -82,7 +82,7 @@ impl FontGlyphRanges {
                 i
             );
             assert!(
-                glyph <= core::char::MAX as u32,
+                glyph <= core::char::MAX as u16,
                 "A glyph in a range cannot exceed the maximum codepoint. \
                  (Glyph is {:#x} at index {})",
                 glyph,
@@ -126,7 +126,7 @@ impl FontGlyphRanges {
     /// # Safety
     ///
     /// It is up to the caller to guarantee the slice contents are valid.
-    pub unsafe fn from_slice_unchecked(slice: &'static [u32]) -> FontGlyphRanges {
+    pub unsafe fn from_slice_unchecked(slice: &'static [u16]) -> FontGlyphRanges {
         FontGlyphRanges::from_ptr(slice.as_ptr())
     }
 
@@ -138,7 +138,7 @@ impl FontGlyphRanges {
     ///
     /// It is up to the caller to guarantee the pointer is not null, remains valid forever, and
     /// points to valid data.
-    pub unsafe fn from_ptr(ptr: *const u32) -> FontGlyphRanges {
+    pub unsafe fn from_ptr(ptr: *const u16) -> FontGlyphRanges {
         FontGlyphRanges(FontGlyphRangeData::Custom(ptr))
     }
